Put `BlendMode`s and `CompositeOperation`s into sub-modules.
This helps avoid name clashes.
This also fixes the ambiguity test on Julia v1.6.0-DEV.
