+++ Project.toml
@@ -1,6 +1,6 @@
 name = "ColorBlendModes"
 uuid = "60508b50-96e1-4007-9d6c-f475c410f16b"
-version = "0.2.0"
+version = "0.2.1"
 
 [deps]
 ColorTypes = "3da002f7-5984-5a60-b8a6-cbb66c0b333f"
