+++ src/types.jl
@@ -5,227 +5,6 @@ struct Hue{T<:Real}
     angle::T
 end
 
-"""
-    BlendMode{mode}
-
-A type used for specifying the mixing mode of opaque colors. The `mode` should
-be a symbol.
-"""
-struct BlendMode{mode} end
-
-"""
-    BlendNormal
-
-The destination color is always the source color.
-```
-    Cdest = Csrc
-```
-"""
-const BlendNormal = BlendMode{:normal}()
-
-
-"""
-    BlendMultiply
-
-The source color is multiplied by the backdrop color.
-```
-    Cdest = Cb × Csrc
-```
-"""
-const BlendMultiply = BlendMode{:multiply}()
-
-
-"""
-    BlendScreen
-
-The complementary colors of the source and backdrop colors are multiplied, and
-the destination color is the complementary color of the multiplicated color.
-```
-    Cdest = 1 - ((1 - Cb) × (1 - Csrc))
-```
-"""
-const BlendScreen = BlendMode{:screen}()
-
-
-"""
-    BlendOverlay
-
-This mode uses the [`multiply`](@ref BlendMultiply) or [`screen`](@ref
-BlendScreen) mode, depending on the backdrop color. The overlay mode is the
-inverse of the [`hard-light`](@ref BlendHardLight) mode.
-```
-    if Cb <= 0.5
-        Cdest = Csrc × 2Cb
-    else
-        Cdest = 1 - ((1 - Csrc) × (1 - 2Cb))
-    end
-```
-"""
-const BlendOverlay = BlendMode{:overlay}()
-
-
-"""
-    BlendDarken
-
-The darker values of the backdrop color and the source color are selected.
-```
-    Cdest = min(Cb, Csrc)
-```
-"""
-const BlendDarken = BlendMode{:darken}()
-
-
-"""
-    BlendLighten
-
-The lighter values of the backdrop color and the source color are selected.
-```
-    Cdest = max(Cb, Csrc)
-```
-"""
-const BlendLighten = BlendMode{:lighten}()
-
-
-"""
-    BlendColorDodge
-
-The destination color is the result of dividing the backdrop color by the
-complementary color of the source color.
-```
-    if Cb == 0
-        Cdest = 0
-    elseif Csrc == 1
-        Cdest = 1
-    else
-        Cdest = min(1, Cb / (1 - Csrc))
-    end
-```
-"""
-const BlendColorDodge = BlendMode{Symbol("color-dodge")}()
-
-
-"""
-    BlendColorBurn
-
-The destination color is the result of dividing the complementary color of the
-backdrop color by the source color.
-
-ColorBlendModes uses the definition of W3C drafts as shown below. Note that
-there is a variant, which returns `0` when `Cb == 1` and `Csrc == 0`.
-```
-    if Cb == 1
-        Cdest = 1
-    elseif Csrc == 0
-        Cdest = 0
-    else
-        Cdest = 1 - min(1, (1 - Cb) / Csrc)
-    end
-```
-"""
-const BlendColorBurn = BlendMode{Symbol("color-burn")}()
-
-
-"""
-    BlendHardLight
-
-This mode uses the [`multiply`](@ref BlendMultiply) or [`screen`](@ref
-BlendScreen) mode, depending on the source color. The overlay mode is the
-inverse of the [`overlay`](@ref BlendOverlay) mode.
-```
-    if Csrc <= 0.5
-        Cdest = Cb × 2Csrc
-    else
-        Cdest = 1 - ((1 - Cb) × (1 - 2Csrc))
-    end
-```
-"""
-const BlendHardLight = BlendMode{Symbol("hard-light")}()
-
-
-"""
-    BlendSoftLight
-
-The result is similar to the [`hard-light`](@ref BlendHardLight) mode, but
-milder. This mode is also related to the [`overlay`](@ref BlendOverlay) mode.
-
-ColorBlendModes uses the definition of W3C drafts as shown below. Note that
-there are different definitions of the `soft-light`.
-```
-    if Csrc <= 0.5
-        Cdest = Cb - (1 - 2Csrc) × Cb × (1 - Cb)
-    else
-        if Cb <= 0.25
-            D = ((4Cb - 3) × Cb + 1) × 4Cb
-        else
-            D = sqrt(Cb)
-        end
-        Cdest = Cb + (2Csrc - 1) × (D - Cb)
-    end
-```
-"""
-const BlendSoftLight = BlendMode{Symbol("soft-light")}()
-
-
-"""
-    BlendDifference
-
-The destination values are the subtraction of the darker values from the lighter
-values of the backdrop and source colors.
-```
-    Cdest = abs(Csrc - Cb)
-```
-"""
-const BlendDifference = BlendMode{:difference}()
-
-
-"""
-    BlendExclusion
-
-The result is similar to the [`difference`](@ref BlendDifference) mode, but
-milder.
-```
-    Cdest = Cb + Csrc - 2Cb × Csrc
-```
-"""
-const BlendExclusion = BlendMode{:exclusion}()
-
-
-"""
-    BlendHue
-
-The result is a color with the hue of the source color and the saturation and
-luminosity of the backdrop color.
-"""
-const BlendHue= BlendMode{:hue}()
-
-
-"""
-    BlendSaturation
-
-The result is a color with the saturation of the source color and the hue and
-luminosity of the backdrop color.
-"""
-const BlendSaturation = BlendMode{:saturation}()
-
-
-"""
-    BlendColor
-
-The result is a color with the hue and saturation of the source color and the
-luminosity of the backdrop color.
-"""
-const BlendColor = BlendMode{:color}()
-
-
-"""
-    BlendLuminosity
-
-The result is a color with the luminosity of the source color and the hue and
-saturation of the backdrop color.
-"""
-const BlendLuminosity = BlendMode{:luminosity}()
-
-
 const SeparableBlendMode = Union{
     BlendMode{:normal},
     BlendMode{:multiply},
@@ -247,125 +26,3 @@ const NonSeparableBlendMode = Union{
     BlendMode{:color},
     BlendMode{:luminosity},
 }
-
-"""
-    CompositeOperation{op}
-
-A type used for specifying the Porter-Duff operator, or the mode of generalized
-alpha compositing. The `op` should be a symbol.
-"""
-struct CompositeOperation{op} end
-
-
-"""
-    CompositeClear
-
-A basic Porter-Duff operator with the fractional terms `Fa = 0; Fb = 0`.
-The composite result is completely transparent.
-"""
-const CompositeClear = CompositeOperation{:clear}()
-
-
-"""
-    CompositeCopy
-
-A basic Porter-Duff operator with the fractional terms `Fa = 1; Fb = 0`.
-The composite result is the copy of source.
-
-!!! note
-    Even if the source alpha is zero, the color components are not cleared.
-"""
-const CompositeCopy = CompositeOperation{:copy}()
-
-
-"""
-    CompositeDestination
-
-A basic Porter-Duff operator with the fractional terms `Fa = 0; Fb = 1`.
-The composite result is the copy of destination (backdrop).
-
-!!! note
-    Even if the destination alpha is zero, the color components are not cleared.
-"""
-const CompositeDestination = CompositeOperation{:destination}()
-
-
-"""
-    CompositeSourceOver
-
-A basic Porter-Duff operator with the fractional terms `Fa = 1; Fb = 1 - αsrc`.
-This means the simple alpha compositing.
-"""
-const CompositeSourceOver = CompositeOperation{Symbol("source-over")}()
-
-
-"""
-    CompositeDestinationOver
-
-A basic Porter-Duff operator with the fractional terms `Fa = 1 - αb; Fb = 1`.
-"""
-const CompositeDestinationOver = CompositeOperation{Symbol("destination-over")}()
-
-
-"""
-    CompositeSourceIn
-
-A basic Porter-Duff operator with the fractional terms `Fa = αb; Fb = 0`.
-"""
-const CompositeSourceIn = CompositeOperation{Symbol("source-in")}()
-
-
-"""
-    CompositeDestinationIn
-
-A basic Porter-Duff operator with the fractional terms `Fa = 0; Fb = αsrc`.
-"""
-const CompositeDestinationIn = CompositeOperation{Symbol("destination-in")}()
-
-
-"""
-    CompositeSourceOut
-
-A basic Porter-Duff operator with the fractional terms `Fa = 1 - αb; Fb = 0`.
-"""
-const CompositeSourceOut = CompositeOperation{Symbol("source-out")}()
-
-
-"""
-    CompositeDestinationOut
-
-A basic Porter-Duff operator with the fractional terms `Fa = 0; Fb = 1 - αsrc`.
-"""
-const CompositeDestinationOut = CompositeOperation{Symbol("destination-out")}()
-
-
-"""
-    CompositeSourceAtop
-
-A basic Porter-Duff operator with the fractional terms `Fa = αb; Fb = 1 - αsrc`.
-"""
-const CompositeSourceAtop = CompositeOperation{Symbol("source-atop")}()
-
-
-"""
-    CompositeDestinationAtop
-
-A basic Porter-Duff operator with the fractional terms `Fa = 1 - αsrc; Fb = αb`.
-"""
-const CompositeDestinationAtop = CompositeOperation{Symbol("destination-atop")}()
-
-
-"""
-    CompositeXor
-
-A basic Porter-Duff operator with the fractional terms `Fa = 1 - αb; Fb = 1 - αsrc`.
-"""
-const CompositeXor = CompositeOperation{:xor}()
-
-
-"""
-    CompositeLighter
-
-A basic Porter-Duff operator with the fractional terms `Fa = 1; Fb = 1`.
-"""
-const CompositeLighter = CompositeOperation{:lighter}()
